Lade till cookies

+++ code/functions.rb
@@ -2,15 +2,20 @@ def checkDatabase(user:, pass:)
   hash = Hash.new
   hashPass = encode(msg: pass)
 
-  File.open("database/database.txt", 'r') do |file|
-    lines = []
-    file.readlines().each do |line|
-      lines << line.chomp
-    end
+  if File.exists?("database/database.txt")
+    File.open("database/database.txt", 'r') do |file|
+      lines = []
+      file.readlines().each do |line|
+        lines << line.chomp
+      end
 
-    lines.each do |line|
-      hash[line[0, line.index(':')]] = line[line.index(':') + 1, line.length]
+      lines.each do |line|
+        hash[line[0, line.index(':')]] = line[line.index(':') + 1, line.length]
+      end
     end
+  else
+    p "Couldn't find database file!"
+    return false
   end
 
   if hash.has_key?(user) && hash[user] == hashPass
